Initial Sentiment Analyser Code
This simply runs VADER, TextBlob and Flair on each sentence from each piece of text in a corpus and puts the results into a pandas dataframe.

Flair slows this program down significantly and VADER could potentially be customised to improve its accuracy.

Need to make sense of this data and potentially store output in csv and potentially create visualisations.