Created Web Scraper for The Independent
Created a new class for scraping articles from The Independent.

Also moved the `update_date_range()` logic to the WebScraper class as this seems to be standard across both sites.

Also moved most of the `WebScraper.scrape()` logic into a try statement with only a finally block, to ensure a timestamp is created even if article retrieval is interrupted.