Feature extraction for unsupervised sentiment analysis
Have created a Python program to extract feature words and bigrams from the labeled corpus and store them in text files, meaning they will not need to be extracted each time the supervised model is ran.

all words are stored lowercase and have had stop words removed.

The NLTK stopwords file has been modified for this, removing negation words as to not impact their relevance to the bias.