Create unsupervised_sa.py
Built a Python program to extract features from a labeled dataset and then use these to train classification models that can be used to identify whether a news article is biased or not.

A range of models are tested here to assess their accuracies so that the best model can be used in the end, or potentially the best few in a majority selection algorithm.