Small changes to WebScraper classes
The `page_is_irrelevant()` method logic has been moved from `BBCWebScraper` to the parent `WebScraper` class as currently this simply ignores articles that are shorter than 5 lines, which is appropriate for all articles.

There is also a small format change to the error output to work better with the `tqdm` progress bars.