Creation of Majority Vote Classifier
The majority vote classifier simply holds 3 classifiers and then outputs the predominant prediction of all 3.

The classifiers used for this are trained and validated on 90% of the dataset using Stratified K Fold and the best of each classifier in this test is then tested again on unseen data and the 3 models with the highest accuracy on the unseen data are used in the MajorityVoteClassifier.

The MajorityVoteClassifier is then also tested on the unseen data to determine the accuracy of this approach.